Backend/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
google-generativeai==0.7.2
pydantic==1.10.11
pillow==10.4.0